Change code to include even more OOP than before.
Let's
*add ui, storage, parser, and tasklist as a means of abstraction
