+++ src/main/java/duke/TaskList.java
@@ -1,0 +1,8 @@
+package duke;
+
+import java.sql.Array;
+import java.util.ArrayList;
+
+public class TaskList extends ArrayList<Task> {
+
+}
