+++ src/main/java/duke/data/duke.txt
@@ -1,1 +1,1 @@
-event cry /from 2007-12-03T15:30 /to 2007-12-03T16:00
+deadline dance /by LOL
